Fase H: correct NSUserDefaults integer bridge types

+++ shared/src/iosMain/kotlin/es/criosrango/shared/IosPendingPaymentStore.kt
@@ -9,10 +9,10 @@ class IosPendingPaymentStore : PendingPaymentStore {
     override fun load(): PendingPayment? {
         val id = defaults.integerForKey("criosrango.payment.order_id")
         val key = defaults.stringForKey("criosrango.payment.order_key")
-        return if (id > 0 && !key.isNullOrBlank()) PendingPayment(id, key) else null
+        return if (id > 0 && !key.isNullOrBlank()) PendingPayment(id.toInt(), key) else null
     }
     override fun save(payment: PendingPayment) {
-        defaults.setInteger(payment.orderId, forKey = "criosrango.payment.order_id")
+        defaults.setInteger(payment.orderId.toLong(), forKey = "criosrango.payment.order_id")
         defaults.setObject(payment.orderKey, forKey = "criosrango.payment.order_key")
     }
     override fun clear() {
